Refactor presenter and synclifecyclecallback classes to 2 presenters
