+++ p2p-sync/src/main/java/org/smartregister/p2p/authenticator/ReceiverConnectionAuthenticator.java
@@ -15,8 +15,8 @@ public class ReceiverConnectionAuthenticator extends BaseSyncConnectionAuthentic
 
     public ReceiverConnectionAuthenticator(@NonNull P2pModeSelectContract.View view
             , @NonNull P2pModeSelectContract.Interactor interactor
-            , @NonNull P2pModeSelectContract.Presenter presenter) {
-        super(view, interactor, presenter);
+            , @NonNull P2pModeSelectContract.BasePresenter basePresenter) {
+        super(view, interactor, basePresenter);
     }
 
     @Override
