+++ p2p-sync/src/main/java/org/smartregister/p2p/contract/BaseView.java
@@ -13,7 +13,7 @@ import org.smartregister.p2p.handler.ActivityResumeHandler;
 
 public interface BaseView extends ActivityResultHandler, ActivityResumeHandler, ActivityRequestPermissionResultHandler {
 
-    void initializePresenter();
+    void initializePresenters();
 
     void runOnUiThread(@NonNull Runnable runnable);
